Alteração nomenclatura studentsRouter

+++ routes/students.js
@@ -5,18 +5,17 @@ const { Router } = require("express");
 
 const StudentsModel = require("../models/StudentsModel.js");
 
-const Students = Router();
+const StudentsRouter = Router();
 
-Students.post("/createStudent", (req, res) => {
+StudentsRouter.post("/createStudent", (req, res) => {
+    let { student_cpf } = req.body;
 
-    let studentInfo = req.body;
-
-    let verification = fieldsVerification("POST", studentInfo, StudentsModel.getAttributes());
+    let verification = fieldsVerification("POST", req.body, StudentsModel.getAttributes());
 
     if (verification.error) {
         res.status(500).json(verification);
     } else {
-        StudentsModel.findOrCreate({ where: { student_cpf: studentInfo.student_cpf }, defaults: studentInfo })
+        StudentsModel.findOrCreate({ where: { student_cpf }, defaults: req.body })
             .then(studentData => {
                 let [studentInformation, created] = studentData;
                 res.status(201).json({ studentInformation, created });
@@ -28,7 +27,7 @@ Students.post("/createStudent", (req, res) => {
 
 });
 
-Students.get("/getStudents", (req, res) => {
+StudentsRouter.get("/getStudents", (req, res) => {
     StudentsModel.findAll()
         .then(studentsData => {
             if (studentsData.length >= 1) {
@@ -39,37 +38,41 @@ Students.get("/getStudents", (req, res) => {
         })
 });
 
-Students.put("/editStudent", (req, res) => {
+StudentsRouter.put("/editStudent", (req, res) => {
 
+    let { student_id } = req.body;
     let infoToEdit = req.body;
-    let { student_id } = infoToEdit;
 
     if (student_id) {
         delete infoToEdit.student_id;
-        
+
         let verification = fieldsVerification("PUT", infoToEdit, StudentsModel.getAttributes());
         if (verification.error) {
             res.status(500).json(verification)
         } else {
-            StudentsModel.update(infoToEdit, { where: { student_id } }).then( studentData => {
-                res.status(200).json({ done: true });
-            }).catch( err => {
-                res.status(500).json({ done: false, err })
+            StudentsModel.update(infoToEdit, { where: { student_id } }).then(_ => {
+                res.status(200).json({ "done": true });
+            }).catch(err => {
+                res.status(500).json({ "done": false, err })
             });
         }
 
     } else {
-        res.status(500).json({ "error": "student_id required" })
+        res.status(400).json({ "error": "student_id required" })
     }
 
 })
 
-Students.delete("/deleteStudent", (req, res) => {
+StudentsRouter.delete("/deleteStudent", (req, res) => {
     let { student_id } = req.body;
 
-    StudentsModel.destroy({ where: { student_id } }).then( wasDeleted => {
-        res.status(200).json({ wasDeleted: !!wasDeleted });
-    })
+    if (student_id) {
+        StudentsModel.destroy({ where: { student_id } }).then(wasDeleted => {
+            res.status(200).json({ wasDeleted: !!wasDeleted });
+        })
+    } else {
+        res.status(400).json({ "error": "student_id required" });
+    }
 });
 
-module.exports = Students;
+module.exports = StudentsRouter;
